+++ src/HoneypotServiceProvider.php
@@ -15,25 +15,24 @@ class HoneypotServiceProvider extends ServiceProvider
         );
     }
 
+    /**
+     * Register the <x-honeypot /> component and the publishable views, translations and config.
+     */
     public function boot(): void
     {
         $this->loadViewsFrom(__DIR__.'/../resources/views', 'livewire-honeypot');
         $this->loadTranslationsFrom(__DIR__.'/../resources/lang', 'livewire-honeypot');
 
-        // Register <x-honeypot />
         Blade::component('livewire-honeypot::components.honeypot', 'honeypot');
 
-        // Allow publishing the views
         $this->publishes([
             __DIR__.'/../resources/views' => resource_path('views/vendor/livewire-honeypot'),
         ], 'livewire-honeypot-views');
 
-        // Allow publishing the translations
         $this->publishes([
             __DIR__.'/../resources/lang' => lang_path('vendor/livewire-honeypot'),
         ], 'livewire-honeypot-translations');
 
-        // Allow publishing the config
         $this->publishes([
             __DIR__.'/../config/livewire-honeypot.php' => config_path('livewire-honeypot.php'),
         ], 'livewire-honeypot-config');